+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2024-09-19
+
+### Fixed
+
+- Fixed the normalization of the `exclude` option when the path is `.` or `./`.
+- Fixed the edge case where the root directory is ignored by the `.gitignore` when the path is `.` or `./` and the `gitignore` has `.*` as an ignore pattern.
+
 ## [0.2.2] - 2024-09-18
 
 ### Added